requirements.txt
plotly == 6.0.0
python-dotenv == 1.0.1
google-generativeai == 0.8.4
streamlit == 1.42.1
pillow == 10.4.0
